pyproject.toml
[tool.poetry]
name = "my-blog"
version = "0.1.0"
description = ""
authors = ["agvantibo <apicalium@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.0.4"
PyYAML = "^6.0"
Markdown = "^3.3.6"
htmlmin = "^0.1.12"
0html = "^0.0.0"
html5print = "^0.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
